test(ui): add MainActivityTest, introduce testTag, and relocate TestRunner
- Add MainActivityTest to verify that the UniversityApp in main activity is displayed correctly
- Introduce testTag in UniversityApp for consistent UI component identification in tests
- Move TestRunner from core.testing to core.testing.util for better organization and utility management
