+++ app/src/androidTest/java/com/kitabisa/test/universitydomains/MainActivityTest.kt
@@ -1,0 +1,33 @@
+package com.kitabisa.test.universitydomains
+
+import androidx.compose.ui.test.assertIsDisplayed
+import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import androidx.compose.ui.test.onNodeWithTag
+import com.kitabisa.test.universitydomains.core.testing.constant.TestTag
+import dagger.hilt.android.testing.HiltAndroidRule
+import dagger.hilt.android.testing.HiltAndroidTest
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+
+@HiltAndroidTest
+class MainActivityTest {
+
+    @get:Rule(order = 0)
+    val hiltRule = HiltAndroidRule(this)
+
+    @get:Rule(order = 1)
+    val composeTestRule = createAndroidComposeRule<MainActivity>()
+
+    @Before
+    fun init() {
+        hiltRule.inject()
+    }
+
+    @Test
+    fun appAreDisplayed() {
+        composeTestRule
+            .onNodeWithTag(TestTag.UNIVERSITY_APP)
+            .assertIsDisplayed()
+    }
+}
